wip: untested solution for clean exit
Ctrl+C on main Broker distributes KILL signals to other Brokers and each
to its respective Nodes.
Brokers continue to broker packets until it encounters an END message
in a stream. It removes tracking for that Node if it's in its domain
and releases it for safe exit. When a Broker no longer has Nodes in its
domain, it requests permission to exit from upstream Brokers, if any.

This should facilitate no loss to uncommitted data still on-air and
reliable running of the program.